requirements.txt
algosdk==2.3.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
stripe==7.6.0
requests==2.31.0
python-multipart==0.0.6 